extract constants
