+++ src/constants/gameData.ts
@@ -1,0 +1,4 @@
+export const fps = 16;
+export const baseSize = 64;
+export const battleSpriteSize = baseSize * 2;
+export const animationTimer = 16000 / fps;
